Create qr_code_reader
You are to write and test a Python/OpenCV program that will perform the first steps of QR code recognition –
that is, to find the QR code in an image and produce a rectified image of the code alone. Notice that a QR code
is delineated by three corner markers that look like this:
To take any general image containing a QR code, you must find three occurrences of this marker in the image,
define an affine mapping to transform the image into a new image that is the proper size and rotation, and
transform the image. So, for an input image that looks like the image on the left, I want an output image that
looks like the one on the right (which is 300 by 300 pixels in size):
I have given you a set of five QR code images. Your program should do the following:
1. Create a suitable template for use in template matching. This should be an 8-bit grayscale image of just
the marker, 128 by 128 pixels in size.
2. For QR_A through QR_B (save QR_E for a bit later):
a. Use cv2.matchTemplate() to find the three locations of the markers in each input image. Note
that it takes a bit of work to examine the output of matchTemplate to find the three best marker
locations!
b. Print to the console the file name and the found locations of the three markers.
c. Write the results array returned by matchTemplate() to disk as a cv2 image.